rename app header and sidebar components
